ABORT DIRECTION
The tests are passing, but the cure seems worse than the disease
here. The goal was to avoid deadlocks but the result is that most
code needs to be written synchronously or else it will get
reordered. I'm not sure the deadlocks are so bad.

I guess that if we wanted to go this direction would just remove
async handlers entirely and encourage everyone to write code
synchronously, and then have some kind of "spawn" for async logic,
so at least you opt-in to the reordering in a very clear way.